Change entity id fields to be Long instead of Int

+++ app/src/main/java/com/example/bookreadingapp/data/entities/Image.kt
@@ -21,7 +21,7 @@ class Image {
     @ColumnInfo(name = "imageId")
     var id: Int = 0
 
-    var chapterId: Int = 0
+    var chapterId: Long = 0
 
     @ColumnInfo(name = "imagePosition")
     var position: Int = 0
